Multiple Revisions
Added Backgrounds to the piece to gain a better sense of place.

Revised multiple parts of the image to make it cleaner, such as adding borders between the two halves of the piece, and adding a white space behind the text.

Added Background Images to Assets

Edited Assets list and copyright therein.

Added two screenshots pertaining to the revisions done.